+++ src/utils/ClonePlus.sol
@@ -4,9 +4,6 @@ pragma solidity ^0.8.4;
 import {Clone} from "./Clone.sol";
 import {IERC20} from "oz/token/ERC20/utils/SafeERC20.sol";
 
-// TODO: Remove import
-import {console} from "forge-std/console.sol";
-
 /// @title ClonePlus
 /// @notice Extends Clone with additional helper functions
 contract ClonePlus is Clone {
@@ -17,7 +14,7 @@ contract ClonePlus is Clone {
     /// @param argOffset The offset of the arg in the packed data
     /// @param arrLen Number of elements in the array
     /// @return arr The array
-    function _getArgIERC20Array(uint256 argOffset, uint64 arrLen)
+    function _getArgIERC20Array(uint256 argOffset, uint256 arrLen)
         internal
         pure
       returns (IERC20[] memory arr)
@@ -39,7 +36,7 @@ contract ClonePlus is Clone {
     /// @param argOffset The offset of the arg in the packed data
     /// @param bytesLen Number of bytes in the data
     /// @return data the bytes data
-    function _getArgBytes(uint256 argOffset, uint64 bytesLen)
+    function _getArgBytes(uint256 argOffset, uint256 bytesLen)
         internal
         pure
       returns (bytes memory data)
